Move dashboards controller to controllers/admin
